pyproject.toml
[tool.poetry]
name = "media-web-dl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.scripts]
media-web-dl = "media_web_dl.__main__:app"

[tool.poetry.dependencies]
python = "^3.10"
certifi = "^2023.7.22"
chardet = "^5.2.0"
charset-normalizer = "^3.3.1"
idna = "^3.4"
nuitka = "^1.8.4"
ordered-set = "^4.1.0"
protobuf = "^3.20.0"
pycryptodome = "^3.19.0"
pycryptodomex = "^3.19.0"
pyperclip = "^1.8.2"
pyyaml = "^6.0.1"
requests = "^2.31.0"
tabulate = "^0.9.0"
urllib3 = "^2.0.7"
wasmer = "^1.1.0"
wasmer-compiler-cranelift = "^1.1.0"
xmltodict = "^0.13.0"
zstandard = "^0.21.0"
typer = {version = "^0.9.0", extras = ["all"]}
colorlog = "^6.7.0"


[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.12"
types-requests = "^2.31.0.10"
types-tabulate = "^0.9.0.3"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
preview = true
line-length = 79
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs        # exclude a few common directories in the
    | \.git         # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py          # also separately exclude a file named foo.py in
                    # the root of the project
)
'''